+++ apps/web/src/app/organizations/tools/import-export/org-import-export.module.ts
@@ -13,7 +13,7 @@ import { ImportService as ImportServiceAbstraction } from "@bitwarden/common/abs
 import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUtils.service";
 import { ImportService } from "@bitwarden/common/services/import.service";
 
-import { LooseComponentsModule } from "../../../modules/loose-components.module";
+import { LooseComponentsModule } from "../../../shared/loose-components.module";
 
 import { OrganizationExportComponent } from "./org-export.component";
 import { OrganizationImportExportRoutingModule } from "./org-import-export-routing.module";
