+++ apps/web/src/app/oss.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from "@angular/core";
 
-import { LooseComponentsModule } from "./modules/loose-components.module";
 import { OrganizationCreateModule } from "./modules/organizations/create/organization-create.module";
 import { OrganizationManageModule } from "./modules/organizations/manage/organization-manage.module";
 import { OrganizationUserModule } from "./modules/organizations/users/organization-user.module";
 import { TrialInitiationModule } from "./modules/trial-initiation/trial-initiation.module";
 import { VaultFilterModule } from "./modules/vault-filter/vault-filter.module";
 import { OrganizationBadgeModule } from "./modules/vault/modules/organization-badge/organization-badge.module";
+import { LooseComponentsModule } from "./shared/loose-components.module";
 import { PipesModule } from "./shared/pipes/pipes.module";
 import { SharedModule } from "./shared/shared.module";
