+++ apps/web/src/app/tools/import-export/import-export.module.ts
@@ -13,7 +13,7 @@ import { ImportService as ImportServiceAbstraction } from "@bitwarden/common/abs
 import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUtils.service";
 import { ImportService } from "@bitwarden/common/services/import.service";
 
-import { LooseComponentsModule } from "../../modules/loose-components.module";
+import { LooseComponentsModule } from "../../shared/loose-components.module";
 
 import { ExportComponent } from "./export.component";
 import { ImportExportRoutingModule } from "./import-export-routing.module";
